列挙体の位置を変更
