+++ entities/constants.go
@@ -15,33 +15,3 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 package entities
-
-type (
-	RunningState int
-	TaskElement  int
-)
-
-const (
-	Main TaskElement = iota
-	StdIn
-	StdOut
-	StdErr
-	numTaskElement
-	NumTaskElement int = int(numTaskElement)
-)
-
-const (
-	Waiting RunningState = iota
-	Running
-	Closed
-	Finalized
-)
-
-func (t TaskElement) String() string {
-	return [...]string{
-		"main task",
-		"standard input",
-		"standard output",
-		"standard error output",
-	}[t]
-}
