+++ entities/element/objects.go
@@ -1,4 +1,4 @@
-// tasks/buffer/packet_base.go
+// entities/element/objects.go
 // Copyright (C) 2021 Kasai Koji
 
 // This program is free software: you can redistribute it and/or modify
@@ -14,21 +14,22 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package buffer
+package element
 
-import (
-	"github.com/rehearsal-open/rehearsal/entities"
-	"github.com/rehearsal-open/rehearsal/entities/element"
+type TaskElement int
+
+const (
+	StdIn TaskElement = iota
+	StdOut
+	StdErr
+	numTaskElement
+	NumTaskElement int = int(numTaskElement)
 )
 
-func (pb *packetBase) Close() error {
-	pb.nClosed++
-	if pb.nClosed >= pb.nSend {
-		pb.bytes = nil
-	}
-	return nil
-}
-
-func (p *packetBase) Sender() (*entities.Task, element.TaskElement) {
-	return p.buffer.task, p.buffer.element
+func (t TaskElement) String() string {
+	return [...]string{
+		"standard input",
+		"standard output",
+		"standard error output",
+	}[t]
 }
